Modify frontend structure
