+++ recommender-front/src/App.jsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { Helmet, HelmetProvider } from 'react-helmet-async';
-import MapComponent from './components/MapComponent';
+import MapComponent from './components/map/MapComponent';
 import 'leaflet/dist/leaflet.css';
+import HeaderComponent from './components/header/HeaderComponent';
 
 function App() {
   return (
@@ -12,12 +13,8 @@ function App() {
           <meta name="description" content="Weather-Based Recommender" />
           <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
         </Helmet>
-        <center>
-          <h1>
-            Weather-Based Recommender
-          </h1>
-          <MapComponent />
-        </center>
+        <HeaderComponent />
+        <MapComponent />
       </div>
     </HelmetProvider>
   );
