+++ recommender-front/src/assets/style.css
@@ -1,0 +1,11 @@
+html, body {
+    width: 100%;
+    height: 100%;  
+    padding: 0;
+    margin: 0;
+  }
+  
+  #map {
+    width: 100vw;
+    height: 100vh;
+  }
