+++ recommender-front/src/components/header/HeaderComponent.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+function HeaderComponent() {
+  return (
+    <center>
+      <h1>
+        Weather-Based Recommender
+      </h1>
+    </center>
+  );
+}
+
+export default HeaderComponent;
