+++ recommender-front/src/components/map/MapComponent.test.jsx
